chatbot/requirements.txt
fastapi
uvicorn
pydantic
requests
qdrant-client
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-llms-openai-like
llama-index-llms-groq
nest-asyncio
python-dotenv
psycopg2-binary
